Release patch version 0.1.2
The GNAT compiler supports filename krunching [sic] for Ada sources,
but there appears to be no simple way to enforce this upon the closure
of sources within a given GPRBuild project. This poses an issue for
library clients that do not use krunching themselves.

One solution would be to expand the filenames, but that would simply
transfer the burden to users of, e.g., GNAT on DJGPP or OpenVMS. As an
alternative, it is possible to explicitly specify a mapping with
GPRBuild's Naming package between Ada specification and body names and
filenames.

* alire.toml: Increment version: 0.1.1 -> 0.1.2.
* LICENSE.txt: rename to license.txt.
* blake2s.gpr:
- Add a license header.
- Minor commenting changes.
- Substitute GPRBuild's Naming package for the '-gnatk8' flag.